fix: NoFilterException is redundant, remove it
In the former version, when collecting filterable 'glob_file'
specs, an unexpected "Traceback" info will be printed in
the debug log.

In fact, it's not necessary to add the NoFilterException,
in that case, the ContentException can be leveraged, we
just need to log it into the debug instead of raising it again.

Signed-off-by: Xiangce Liu <xiangceliu@redhat.com>